@@ -18,7 +18,7 @@
 package org.apache.tvm.tvmrpc;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.widget.Button;
 import android.view.View;
